feat: add exists validation in repository

+++ app/Http/Repositories/TaskRepository.php
@@ -3,6 +3,7 @@
 namespace App\Http\Repositories;
 
 use App\Models\Task;
+use Illuminate\Contracts\Queue\EntityNotFoundException;
 use Illuminate\Database\Eloquent\Collection;
 
 class TaskRepository {
@@ -12,6 +13,9 @@ class TaskRepository {
     }
 
     public function show(int $id): Task {
+        if(!Task::where('id', $id)->exists()) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         return Task::find($id);
     }
 
@@ -26,6 +30,9 @@ class TaskRepository {
     }
 
     public function destroy(int $id): void {
+        if(!Task::where('id', $id)->exists()) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         Task::destroy($id);
     }
